src/main/webapp/app/entities/exit-formalties/route/exit-formalties-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IExitFormalties } from '../exit-formalties.model';
import { ExitFormaltiesService } from '../service/exit-formalties.service';

import { ExitFormaltiesRoutingResolveService } from './exit-formalties-routing-resolve.service';

describe('ExitFormalties routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: ExitFormaltiesRoutingResolveService;
  let service: ExitFormaltiesService;
  let resultExitFormalties: IExitFormalties | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(ExitFormaltiesRoutingResolveService);
    service = TestBed.inject(ExitFormaltiesService);
    resultExitFormalties = undefined;
  });

  describe('resolve', () => {
    it('should return IExitFormalties returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultExitFormalties = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultExitFormalties).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultExitFormalties = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultExitFormalties).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IExitFormalties>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultExitFormalties = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultExitFormalties).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
